+++ UICommon/_events/WasabyEvents.ts
@@ -1,4 +1,4 @@
-import {detection} from 'Env/Env';
+import {constants, detection} from 'Env/Env';
 import {Logger} from 'UICommon/Utils';
 
 import * as EventUtils from './EventUtils';
@@ -7,10 +7,9 @@ import {
     IEventConfig,
     IWasabyEventSystem,
     IHandlerInfo,
-    IArrayEvent
+    IArrayEvent, IClickEvent
 } from './IEvents';
 
-import {TouchHandlers} from './Touch/TouchHandlers';
 import {FastTouchEndController} from './Touch/FastTouchEndController';
 import {SwipeController} from './Touch/SwipeController';
 import {LongTapController} from './Touch/LongTapController';
@@ -29,6 +28,7 @@ import {
  */
 
 const callAfterMount: IArrayEvent[] = [];
+let touchId: number = 0;
 
 abstract class WasabyEvents implements IWasabyEventSystem {
     private capturedEventHandlers: Record<string, IHandlerInfo[]>;
@@ -41,7 +41,17 @@ abstract class WasabyEvents implements IWasabyEventSystem {
     protected _rootDOMNode: TModifyHTMLNode;
     private _handleTabKey: Function;
 
-    protected touchHandlers: TouchHandlers;
+    private clickStateTarget: { target: HTMLElement, touchId: number }[] = [];
+    private preventShouldUseClickByTap: boolean = false;
+    private clickState = {
+        detected: false,
+        stage: '',
+        timer: undefined,
+        timeout: 500,
+        target: null,
+        touchCount: 0,
+        timeStart: undefined
+    };
 
     //#region инициализация системы событий
     protected constructor(rootNode: TModifyHTMLNode, tabKeyHandler?: Function) {
@@ -50,7 +60,6 @@ abstract class WasabyEvents implements IWasabyEventSystem {
         this._handleTabKey = tabKeyHandler;
 
         this.initEventSystemFixes();
-        this.touchHandlers = new TouchHandlers(this._handleClick, this.captureEventHandler);
 
         // если я это не напишу, ts ругнется 'touchendTarget' is declared but its value is never read
         this.touchendTarget = this.touchendTarget || null;
@@ -280,7 +289,7 @@ abstract class WasabyEvents implements IWasabyEventSystem {
 
     //#region специфические обработчики
     protected _handleClick(event: MouseEvent): void {
-        this.touchHandlers.shouldUseClickByTapOnClick(event);
+        this.shouldUseClickByTapOnClick(event);
 
         /**
          * Firefox right click bug
@@ -298,7 +307,9 @@ abstract class WasabyEvents implements IWasabyEventSystem {
 
         // Break click on non-empty selection with type "Range".
         // Have to trim because of fake '\n' selection in some cases.
-        const hasSelection = selection && selection.type === 'Range' && (event.target as HTMLElement).contains(selection.focusNode);
+        const hasSelection = selection &&
+            selection.type === 'Range' &&
+            (event.target as HTMLElement).contains(selection.focusNode);
         const userSelectIsNone = window && window.getComputedStyle
             ? window.getComputedStyle(event.target as HTMLElement)['user-select'] === 'none'
             : true;
@@ -336,7 +347,7 @@ abstract class WasabyEvents implements IWasabyEventSystem {
 
     // TODO: docs
     protected _handleTouchmove(event: ITouchEvent): void {
-        this.touchHandlers.shouldUseClickByTapOnTouchmove(event);
+        this.shouldUseClickByTapOnTouchmove(event);
         FastTouchEndController.setClickEmulateState(false);
         SwipeController.detectState(event);
         LongTapController.resetState();
@@ -344,7 +355,7 @@ abstract class WasabyEvents implements IWasabyEventSystem {
 
     // TODO: docs
     protected _handleTouchend(event: ITouchEvent): void {
-        this.touchHandlers.shouldUseClickByTapOnTouchend(event);
+        this.shouldUseClickByTapOnTouchend(event);
 
         // Compatibility. Touch events handling in Control.compatible looks for
         // the `addedToClickState` flag to see if the event has already been
@@ -372,6 +383,155 @@ abstract class WasabyEvents implements IWasabyEventSystem {
     }
 
     //#endregion
+    //#region обработка тача на специфичных устройствах
+    protected shouldUseClickByTapOnClick(event: MouseEvent): void {
+        if (this.shouldUseClickByTap()) {
+            const idx = this.getClickStateIndexForTarget(this.fixSvgElement(event.target));
+            // if click event occurred, we can remove monitored target
+            if (idx > -1) {
+                this.clickStateTarget.splice(idx, 1);
+            }
+        }
+    }
+
+    protected shouldUseClickByTapOnTouchstart(event: ITouchEvent): void {
+        if (this.shouldUseClickByTap()) {
+            // Для svg запоминаем ownerSVGElement, т.к. иногда в touchstart таргет - это тег svg,
+            // при этом у события click, таргетом будет внутренний элемент
+            const target = this.fixSvgElement(event.target);
+            this.clickStateTarget.push({
+                target,
+                touchId: touchId++ // записываем номер текущего касания
+            });
+        }
+    }
+
+    protected shouldUseClickByTapOnTouchmove(event: ITouchEvent): void {
+        if (this.shouldUseClickByTap()) {
+            this.clickState.touchCount++;
+            // Only one touchmove event is allowed between touchstart and touchend events on Ipad.
+            // If more than one touchmove did occurred, we don't emulate click event.
+            // But on windows installed devices touchmove event can occur some times,
+            // therefore we must check if touchmove count more than 1.
+            if (this.clickState.touchCount > 3) {
+                const idx = this.getClickStateIndexForTarget(this.fixSvgElement(event.target));
+                if (idx > -1) {
+                    this.clickStateTarget.splice(idx, 1);
+                }
+            }
+        }
+    }
+
+    protected shouldUseClickByTapOnTouchend(event: ITouchEvent): void {
+        if (this.shouldUseClickByTap() && !this.preventShouldUseClickByTap) {
+            const lastTouchId = touchId;
+            this.clickState.touchCount = 0;
+            // click occurrence checking
+            setTimeout(() => {
+                // Вызываем клик, если клик был не вызван автоматически после touchEnd. Такое иногда
+                // происходит на тач-телевизорах и планшетах на Windows, и в ограниченной версии
+                // вебкита, используемой например в Presto Offline.
+                // Для того чтобы понять, нужно ли нам эмулировать клик, проверяем два условия:
+                // 1. Элемент, на котором сработал touchEnd, есть в массиве clickStateTarget
+                //    (туда они добавляются при touchStart, и удаляются, если на этом элементе
+                //    срабатывает touchMove или click)
+                // 2. Если этот элемент там есть, проверяем что он соответствует именно тому touchStart,
+                //    который является парным для этого touchEnd. Это можно определить по номеру касания
+                //    touchId. Это предотвращает ситуации, когда мы быстро нажимаем на элемент много
+                //    раз, и этот setTimeout, добавленный на первое касание, находит в массиве clickStateTarget
+                //    тот же элемент, но добавленный на сотое касание.
+                const idx = this.getClickStateIndexForTarget(this.fixSvgElement(event.target));
+                if (idx > -1 && this.clickStateTarget[idx].touchId < lastTouchId) {
+                    // If the click did not occur, we emulate the click through the
+                    // vdom environment only (so that the old WS3 environment ignores it).
+                    // To do so, we generate the fake click event object based on the data
+                    // from the touchend event and propagate it using the vdom bubbling.
+                    const clickEventObject = this.generateClickEventFromTouchend(event) as MouseEvent;
+                    this._handleClick(clickEventObject);
+                    this.captureEventHandler(clickEventObject);
+                }
+            }, this.clickState.timeout);
+        }
+    }
+
+    protected setPreventShouldUseClickByTap(value: boolean): void {
+        this.preventShouldUseClickByTap = value;
+    }
+    /*
+     * Обеспечивает правильную работу тач событий на телевизорах с тачем и windows планшетах
+    */
+    private shouldUseClickByTap(): boolean {
+        // In chrome wrong target comes in event handlers of the click events on touch devices.
+        // It occurs on the TV and the Windows tablet. Presto Offline uses limited version of WebKit
+        // therefore the browser does not always generate clicks on the tap event.
+        return (
+            constants.browser.isDesktop ||
+            (constants.compatibility.touch &&
+                constants.browser.chrome &&
+                navigator &&
+                navigator.userAgent.indexOf('Windows') > -1)
+        );
+    }
+
+    private generateClickEventFromTouchend(event: TouchEvent): IClickEvent {
+        let touch: any = event.changedTouches && event.changedTouches[0];
+        if (!touch) {
+            touch = {
+                clientX: 0,
+                clientY: 0,
+                screenX: 0,
+                screenY: 0
+            };
+        }
+
+        // We do not use document.createEvent or new MouseEvent to make an
+        // actual event object, because in that case we can not change
+        // the target - target property is non-configurable in some
+        // browsers.
+        // We create a simple object instead and fill in the fields we might
+        // need.
+        return {
+            type: 'click',
+            bubbles: event.bubbles,
+            cancelable: event.cancelable,
+            view: window,
+            detail: 1,
+            screenX: touch.screenX,
+            screenY: touch.screenY,
+            clientX: touch.clientX,
+            clientY: touch.clientY,
+            ctrlKey: event.ctrlKey,
+            altKey: event.altKey,
+            shiftKey: event.shiftKey,
+            metaKey: event.metaKey,
+            button: 0,
+            buttons: 0,
+            relatedTarget: null,
+            target: event.target,
+            currentTarget: event.currentTarget,
+            eventPhase: 1, // capture phase
+            stopPropagation(): void {
+                this.bubbles = false;
+            },
+            preventDefault(): void {
+                // no action
+            }
+        };
+    }
+
+    //#endregion
+
+    // Возвращает самое старое (т. к. они расположены по порядку) касание, для которого
+    // сработал touchStart, но для которого не было touchMove или click
+    private getClickStateIndexForTarget(target: HTMLElement): number {
+        return this.clickStateTarget.findIndex((el: any): boolean => el.target === target);
+    }
+
+
+    private fixSvgElement(element: EventTarget): HTMLElement {
+        return (element as SVGElement).ownerSVGElement ?
+            (element as SVGElement).ownerSVGElement as unknown as HTMLElement : element as HTMLElement;
+    }
 
     //#region _notify события
     /**
