+++ UIReact/ReactUnitTest/_events/eventForwarding.test.tsx
@@ -103,32 +103,32 @@ describe('Подписки на контролы', () => {
 
     // TODO: расскоментировать после выполнения (проблема с контекстом)
     // https://online.sbis.ru/opendoc.html?guid=e4cb8aee-57e5-4c8e-9902-b69828cdf5d3
-    // it('Проверяем события тача', async () => {
-    //     global.navigation = { maxTouchPoints: 1 };
-    //     const originalTouchState = eventSystem.touchHandlers.shouldUseClickByTap;
-    //     eventSystem.touchHandlers.shouldUseClickByTap = () => {
-    //         return true;
-    //     };
-    //
-    //     let instance;
-    //     act(() => {
-    //         instance = render(<CounterControl/>, container);
-    //     });
-    //     tick(0);
-    //
-    //     const button = container.querySelector('button');
-    //     assert.equal(instance.clickCount, '0');
-    //
-    //     act(() => {
-    //         button.dispatchEvent(new window.TouchEvent('touchstart', {bubbles: true}));
-    //         button.dispatchEvent(new window.TouchEvent('touchend', {bubbles: true}));
-    //     });
-    //     await tickAsync(500);
-    //     tick(0);
-    //
-    //     assert.equal(instance.clickCount,'1');
-    //
-    //     eventSystem.touchHandlers.shouldUseClickByTap = originalTouchState;
-    //     delete global.navigation;
-    // });
+    it('Проверяем события тача', async () => {
+        global.navigation = { maxTouchPoints: 1 };
+        const originalTouchState = eventSystem.shouldUseClickByTap;
+        eventSystem.shouldUseClickByTap = () => {
+            return true;
+        };
+
+        let instance;
+        act(() => {
+            instance = render(<CounterControl/>, container);
+        });
+        tick(0);
+
+        const button = container.querySelector('button');
+        assert.equal(instance.clickCount, '0');
+
+        act(() => {
+            button.dispatchEvent(new window.TouchEvent('touchstart', {bubbles: true}));
+            button.dispatchEvent(new window.TouchEvent('touchend', {bubbles: true}));
+        });
+        await tickAsync(500);
+        tick(0);
+
+        assert.equal(instance.clickCount,'1');
+
+        eventSystem.shouldUseClickByTap = originalTouchState;
+        delete global.navigation;
+    });
 });
